Refactor AppDelegate layer: single effect funnel, no coordinator indirection
Architectural cleanup with no behavior changes, reviewed end-to-end by a
multi-agent Fable review pass:

- Collapse TrackingRuntimeClient from 18 closures to a single
  perform(EffectIntent) funnel; AppDelegate.performTrackingEffect is now
  the only place reducer effects execute
- Delete the inner TrackingCoordinator class and its ~60 forwarding
  shims; logic lives on AppDelegate split by concern (+Tracking,
  +Calibration, +DeviceEvents)
- Route every store dispatch through applyTrackingAction /
  sendTrackingAction so detector/UI sync can never be skipped
- Unify the duplicated calibration flows into one source-parameterized
  path; fixes a stale calibratingSource leak on detector start failure
- Make the test suite headless: menu construction split from status-item
  creation, integration tests no longer need a window server, new
  MenuBarManagerHeadlessTests restores menu bar update coverage
  (367 -> 404 runnable tests)
- Reorganize Sources/ into Core, AppDelegate, Detectors, System,
  Settings, UI; split SettingsWindow.swift into view and components
- Update CLAUDE.md key files section and add architecture rules
