+++ Sources/AppDelegate/AppDelegate+Overlay.swift
