+++ Sources/Detectors/CameraPostureDetector.swift
