+++ Sources/System/CameraObserver.swift
