+++ Sources/UI/AnalyticsWindow.swift
